Fix thread safety bugs, null dereferences, logic errors, and missing config checks
- Fix Bulwark skip gate using || instead of && in PLD MitigationModule
- Add missing EnableFoul and EnableParadox config toggle checks in BLM DamageModule
- Split shared cleanup buffer across separate locks in DamageIntakeService
- Fix forecast entity buffer race condition in DamageIntakeService
- Fix TankCooldownService peak reset overwriting with low value instead of clearing
- Clamp AetherflowTrackingService cooldown to prevent negative return values
- Add lock-protected read for co-healer entity ID in CoHealerDetectionService
- Add StatusList null guard in DebuffDetectionService
- Add null party member guards in PartyAnalyzerService
- Fix int overflow for boss HP in HpPredictionService by using long cast
- Copy inner lists in ShieldTrackingService snapshot methods
- Floor MP consumption rate window to 1s to prevent burst inflation
- Fix IPC protocol version check to reject messages with missing ver field
- Fix Conservative/Aggressive preset SheltronMinGauge values being inverted
- Fix dispose ordering so event subscribers unsubscribe before container teardown
- Fix test false positive in BuffModuleTests by adding missing TryExecute call
- Update MpForecastService tests for corrected rate floor behavior

+++ Services/Prediction/HpPredictionService.cs
@@ -110,7 +110,7 @@ public sealed class HpPredictionService : IHpPredictionService, IDisposable
             totalPendingHeal = (int)(totalPendingHeal * (1f - varianceReduction));
         }
 
-        return (uint)Math.Clamp(baseHp + totalPendingHeal, 0, (int)maxHp);
+        return (uint)Math.Clamp(baseHp + totalPendingHeal, 0L, (long)maxHp);
     }
 
     /// <summary>
@@ -310,7 +310,7 @@ public sealed class HpPredictionService : IHpPredictionService, IDisposable
         var predictedHp = baseHp + pendingHeals + shieldValue - predictedDamage;
 
         // Clamp to valid range
-        return Math.Clamp(predictedHp, 0, (int)maxHp);
+        return (int)Math.Clamp((long)predictedHp, 0L, (long)maxHp);
     }
 
     /// <summary>
